requirements-pi.txt
opencv-python
pyzmq
msgpack
msgpack-numpy
pigpio
cv2